+++ custom_components/awnet_local/const_sensor.py
@@ -604,6 +604,7 @@ SENSOR_DESCRIPTIONS = (
         name="Lightning Strike Distance",
         icon="mdi:lightning-bolt",
         native_unit_of_measurement=UnitOfLength.KILOMETERS,
+        device_class=SensorDeviceClass.DISTANCE,
         state_class=SensorStateClass.MEASUREMENT,
     ),
     SensorEntityDescription(