+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [Unreleased] - 2023-07-13
+
+### Fixed
+
+- Lightning Strike Distance fixed to be `distance` device class
+- Device class added for evapotranspiration and GDD sensors
+- Housekeeping: updated devcontainer configuation for Python 3.11
+
 ## [1.1.2] - 2023-06-16
 
 ### Fixed